Distribucion del archivo area de trabajo por exceso de codigo, nueva distribucion por modulos para cada panel, inicio de archivo configuracion para cambio de tema de claro a oscuro a futuro, se agrega el menu de figuras con sus respectivos botones
