+++ config.py
@@ -1,0 +1,9 @@
+#Colores modo claro 
+C_BARRA_SUPERIOR_CLARO = "B0BBC6"
+C_BOTON_MENU_CLARO = "#7595B8"
+C_BOTON_AREATRABAJO_CLARO = "DodgerBlue2"
+
+#Colore modo oscurio
+C_BARRA_SUPERIOR_OSCURO = "#2A3138"
+C_BOTON_MENU_OSCURO = "#32404F"
+C_BOTON_AREATRABAJO_OSCURO = "DodgerBlue2"
