+++ src/forms/form_area_trabajo/form_calculadora/calculadora.py
@@ -1,0 +1,6 @@
+#importamos librerias
+import tkinter as tk
+
+def CalculadoraPanel(self):
+    texto = tk.Label(self.AreaTrabajo, text="Calculadora", font = ("Arial", 35, "bold"), bg="light grey")
+    texto.pack(side=tk.TOP, fill="both", expand=True)
