+++ src/forms/form_area_trabajo/form_convercion/conversion.py
@@ -1,0 +1,6 @@
+#importamos librerias
+import tkinter as tk
+
+def conversionNumerica(self):
+    texto = tk.Label(self.AreaTrabajo, text="Conversion numerica", font = ("Arial", 35, "bold"), bg="light grey")
+    texto.pack(side=tk.TOP, fill="both", expand=True)
