+++ src/forms/form_area_trabajo/form_figuras/figuras.py
@@ -1,0 +1,5 @@
+import tkinter as tk
+
+def figurasGeometricas(self):
+    texto = tk.Label(self.AreaTrabajo, text="Figuras geometricas", font = ("Arial", 35, "bold"), bg="light grey")
+    texto.pack(side=tk.TOP, fill="both", expand=True)
